@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "axios": "^0.27.2",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-icons": "^4.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
     "react-scripts": "^5.0.1",
-    "react-test-renderer": "^18.1.0",
+    "react-test-renderer": "^18.2.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -38,8 +38,8 @@
     ]
   },
   "devDependencies": {
-    "@playwright/test": "^1.22.2",
-    "@testing-library/dom": "^8.13.0",
-    "@testing-library/user-event": "^14.1.1"
+    "@playwright/test": "^1.31.2",
+    "@testing-library/dom": "^8.20.0",
+    "@testing-library/user-event": "^14.4.3"
   }
 }
