@@ -1,7 +1,7 @@
 // TODO: Update this, don't want any type
 const pokemonSchema = (table: any) => {
   table.increments("id").primary().unique();
-  table.string("dexID").notNullable();
+  table.integer("dexID").unique().notNullable();
   table.string("name").notNullable();
   table.string("num").notNullable();
   table.string("primaryType").notNullable();
